+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Parm Change Log
 
 * Fix setup.py adding data files
 * Fix heading 1 not found 
+* Add source file view support
+* Add rst transfer to markdown support
+* Add rst2md sub command, it can convert rst to markdown
 
 0.1 Version
 -----------------